Refactor: 디렉토리 구조 수정 변경 전 : domain 변경 후 : layer
