--- src/main/java/com/numble/popularpuppyvote/common/QueryDslConfig.java
@@ -1,21 +1,0 @@
-package com.numble.popularpuppyvote.common;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-@Configuration
-public class QueryDslConfig {
-
-	@PersistenceContext
-	EntityManager entityManager;
-
-	@Bean
-	public JPAQueryFactory jpaQueryFactory() {
-		return new JPAQueryFactory(entityManager);
-	}
-}
