+++ src/main/java/com/numble/popularpuppyvote/domain/dto/request/LikesRegisterRequest.java
@@ -1,0 +1,7 @@
+package com.numble.popularpuppyvote.domain.dto.request;
+
+public record LikesRegisterRequest(
+		Long puppyId
+) {
+
+}
