+++ src/main/java/com/numble/popularpuppyvote/domain/dto/request/PuppyCreateRequest.java
@@ -1,0 +1,9 @@
+package com.numble.popularpuppyvote.domain.dto.request;
+
+public record PuppyCreateRequest (
+		String name,
+		String imageUrl,
+		String description
+) {
+
+}
