+++ src/main/java/com/numble/popularpuppyvote/domain/dto/request/PuppyGetRequest.java
@@ -1,0 +1,6 @@
+package com.numble.popularpuppyvote.domain.dto.request;
+
+public record PuppyGetRequest(
+		Long id
+) {
+}
