+++ src/main/java/com/numble/popularpuppyvote/domain/dto/request/PuppyListGetRequest.java
@@ -1,0 +1,7 @@
+package com.numble.popularpuppyvote.domain.dto.request;
+
+public record PuppyListGetRequest(
+		Long cursorId,
+		int pageSize
+) {
+}
