+++ src/main/java/com/numble/popularpuppyvote/domain/dto/request/PuppyUpdateRequest.java
@@ -1,0 +1,8 @@
+package com.numble.popularpuppyvote.domain.dto.request;
+
+public record PuppyUpdateRequest(
+		Long id,
+		String imageUrl,
+		String description
+) {
+}
