+++ src/main/java/com/numble/popularpuppyvote/domain/dto/response/LikesRegisterResponse.java
@@ -1,0 +1,6 @@
+package com.numble.popularpuppyvote.domain.dto.response;
+
+public record LikesRegisterResponse(
+		Long id
+) {
+}
