+++ src/main/java/com/numble/popularpuppyvote/domain/dto/response/PuppyGetResponse.java
@@ -1,0 +1,10 @@
+package com.numble.popularpuppyvote.domain.dto.response;
+
+public record PuppyGetResponse(
+		Long id,
+		String name,
+		String imageUrl,
+		String description,
+		int likeCount
+) {
+}
