+++ src/main/java/com/numble/popularpuppyvote/domain/dto/response/PuppyListGetResponse.java
@@ -1,0 +1,9 @@
+package com.numble.popularpuppyvote.domain.dto.response;
+
+import java.util.List;
+
+public record PuppyListGetResponse(
+		List<PuppyGetResponse> puppies,
+		Long lastId
+) {
+}
