+++ src/main/java/com/numble/popularpuppyvote/domain/dto/response/PuppyUpdateResponse.java
@@ -1,0 +1,6 @@
+package com.numble.popularpuppyvote.domain.dto.response;
+
+public record PuppyUpdateResponse(
+		Long id
+) {
+}
