--- src/main/java/com/numble/popularpuppyvote/domain/likes/controller/LikesController.java
@@ -1,37 +1,0 @@
-package com.numble.popularpuppyvote.domain.likes.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.numble.popularpuppyvote.domain.likes.dto.request.LikesRegisterRequest;
-import com.numble.popularpuppyvote.domain.likes.dto.response.LikesRegisterResponse;
-import com.numble.popularpuppyvote.domain.likes.service.LikesService;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/api/v1/likes")
-@RequiredArgsConstructor
-public class LikesController {
-	private final LikesService likesService;
-
-	@PostMapping
-	@ResponseStatus(code = HttpStatus.CREATED)
-	public ResponseEntity<LikesRegisterResponse> registerLikes(@RequestBody LikesRegisterRequest request){
-		return new ResponseEntity<>(likesService.registerLikes(request), HttpStatus.CREATED);
-	}
-
-	@PatchMapping("/{likesId}")
-	@ResponseStatus(code = HttpStatus.OK)
-	public ResponseEntity<Void> deleteLikes(@PathVariable Long likesId){
-		likesService.deleteLikes(likesId);
-		return ResponseEntity.ok().build();
-	}
-}
