--- src/main/java/com/numble/popularpuppyvote/domain/likes/dto/request/LikesRegisterRequest.java
@@ -1,7 +1,0 @@
-package com.numble.popularpuppyvote.domain.likes.dto.request;
-
-public record LikesRegisterRequest(
-		Long puppyId
-) {
-
-}
