--- src/main/java/com/numble/popularpuppyvote/domain/likes/dto/response/LikesRegisterResponse.java
@@ -1,6 +1,0 @@
-package com.numble.popularpuppyvote.domain.likes.dto.response;
-
-public record LikesRegisterResponse(
-		Long id
-) {
-}
