--- src/main/java/com/numble/popularpuppyvote/domain/likes/repository/LikesRepository.java
@@ -1,10 +1,0 @@
-package com.numble.popularpuppyvote.domain.likes.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.numble.popularpuppyvote.domain.likes.model.Likes;
-
-@Repository
-public interface LikesRepository extends JpaRepository<Likes, Long> {
-}
