--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/request/PuppyCreateRequest.java
@@ -1,9 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.request;
-
-public record PuppyCreateRequest (
-		String name,
-		String imageUrl,
-		String description
-) {
-
-}
