--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/request/PuppyGetRequest.java
@@ -1,6 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.request;
-
-public record PuppyGetRequest(
-		Long id
-) {
-}
