--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/request/PuppyListGetRequest.java
@@ -1,7 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.request;
-
-public record PuppyListGetRequest(
-		Long cursorId,
-		int pageSize
-) {
-}
