--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/request/PuppyUpdateRequest.java
@@ -1,8 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.request;
-
-public record PuppyUpdateRequest(
-		Long id,
-		String imageUrl,
-		String description
-) {
-}
