--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/response/PuppyCreateResponse.java
@@ -1,6 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.response;
-
-public record PuppyCreateResponse(
-		Long id
-) {
-}
