--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/response/PuppyGetResponse.java
@@ -1,10 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.response;
-
-public record PuppyGetResponse(
-		Long id,
-		String name,
-		String imageUrl,
-		String description,
-		int likeCount
-) {
-}
