--- src/main/java/com/numble/popularpuppyvote/domain/puppy/dto/response/PuppyListGetResponse.java
@@ -1,9 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.dto.response;
-
-import java.util.List;
-
-public record PuppyListGetResponse(
-		List<PuppyGetResponse> puppies,
-		Long lastId
-) {
-}
