--- src/main/java/com/numble/popularpuppyvote/domain/puppy/repository/PuppyCustomRepository.java
@@ -1,10 +1,0 @@
-package com.numble.popularpuppyvote.domain.puppy.repository;
-
-import java.util.List;
-
-import com.numble.popularpuppyvote.domain.puppy.model.Puppy;
-
-public interface PuppyCustomRepository {
-
-	List<Puppy> findPuppies(Long cursorId, int pageSize);
-}
