+++ src/main/java/com/numble/popularpuppyvote/domain/repository/PuppyCustomRepository.java
@@ -1,0 +1,10 @@
+package com.numble.popularpuppyvote.domain.repository;
+
+import java.util.List;
+
+import com.numble.popularpuppyvote.domain.model.Puppy;
+
+public interface PuppyCustomRepository {
+
+	List<Puppy> findPuppies(Long cursorId, int pageSize);
+}
